backend/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.2
exceptiongroup==1.2.0
fastapi==0.109.2
greenlet==3.0.3
h11==0.14.0
idna==3.6
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
PyMySQL==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.8
sniffio==1.3.0
SQLAlchemy==2.0.25
SQLAlchemy-Utils==0.41.1
starlette==0.36.3
typing_extensions==4.9.0
uvicorn==0.27.1
